# backend/swagger.yaml
swagger: "2.0"
info:
  description: "A web-based  judging  application  that  allows  users,  namely Admin, to host events and Participants, to create their project ideas or join an existing project and Judges, to be seamlessly involved in rating the assigned projects. Participants create and join the projects in an event. Eventually, judges are allowed to rate and admins can announce the winners of the projects in each event that are assigned to them. Each user has their roles and access to perform their respective operations described in the upcoming sections."
  version: "1.0.0"
  title: "SeamlessJudging"
host: "stark.cse.buffalo:3000"
schemes:
- "http"
paths:
...
  /api/v1/course-code:
    get:
      tags:
      - "CourseCode"
      summary: "Retrieve Course Code Details"
      description: "Only admin retreive course-codes complete list"
      operationId: "getCourseCodes"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      responses:
          "200":
            description: "successful"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Course-Code details retrieved successfully"
                response_data:
                  type: "array"
                  items:
                    $ref: "#/definitions/CourseCode"
          "404":
            description: "Not Found"

    post:
      tags:
      - "CourseCode"
      summary: "Add a CourseCode"
      description: "Only admin can add a CourseCode"
      operationId: "addCourseCode"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Add a CourseCode"
        required: true
        schema:
          $ref: "#/definitions/AddCourseCode"
      responses:
          "201":
            description: "successfully added"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "CourseCode added succesfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
    
    delete:
      tags:
      - "CourseCode"
      summary: "Delete course-codes in bulk"
      description: "Delete course-codes in bulk"
      operationId: "deleteCourseCodes"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to delete course-codes"
        required: true
        schema:
          type: "object"
          properties:
            ids:
              $ref: "#/definitions/ArrayOfInt"
      responses:
          "201":
            description: "Course-codes deleted successfully"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "Course-codes deleted successfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
  /api/v1/course-code/{courseCodeId}:
    put:
      tags:
      - "CourseCode"
      summary: "Update a CourseCode"
      description: "Only admin can update a CourseCode"
      operationId: "updateCourseCode"
      produces:
      - "application/json"
      parameters:
      - name: "Access-Token"
        in: "header"
        description: ""
        required: true
        type: "string"
      - name: "courseCodeId"
        in: "path"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request to update a CourseCode"
        required: true
        schema:
          $ref: "#/definitions/AddCourseCode"
      responses:
          "200":
            description: "successfully updated"
            schema:
              type: "object"
              properties:
                response_str:
                    type: "string"
                    example: "CourseCode updated succesfully"
                response_data:
                  type: "object"
          "404":
            description: "Not Found"
...
definitions:
...
  ArrayOfInt:
    type: array
    items:
      type: integer
...
  CourseCode:
    type: "object"
    properties:
      course_code_id:
          type: integer
          example: 1
      name:
          type: string
          example: "MS Project Development"
      code:
          type: string
          example: "CSE-611"
  AddCourseCode:
    type: "object"
    properties:
      name:
          type: string
          example: "MS Project Development"
      code:
          type: string
          example: "CSE-611"
...